+++ ElvUI_LuckyoneUI/Profiles/Quartz.lua
@@ -1,5 +1,4 @@
 local L1UI, E, L, V, P, G = unpack(select(2, ...))
-if not E.Retail then return end
 
 local pairs = pairs